+++ ignite/cmd/scaffold_chain.go
@@ -83,6 +83,7 @@ about Cosmos SDK on https://docs.cosmos.network
 	c.Flags().Bool(flagNoDefaultModule, false, "create a project without a default module")
 	c.Flags().StringSlice(flagParams, []string{}, "add default module parameters")
 	c.Flags().Bool(flagSkipGit, false, "skip Git repository initialization")
+	c.Flags().Bool(flagSkipProto, false, "skip proto generation")
 	c.Flags().Bool(flagMinimal, false, "create a minimal blockchain (with the minimum required Cosmos SDK modules)")
 	c.Flags().Bool(flagIsConsumer, false, "scafffold an ICS consumer chain")
 	// Cannot have both minimal and consumer flag
@@ -104,6 +105,7 @@ func scaffoldChainHandler(cmd *cobra.Command, args []string) error {
 		minimal, _         = cmd.Flags().GetBool(flagMinimal)
 		isConsumer, _      = cmd.Flags().GetBool(flagIsConsumer)
 		params, _          = cmd.Flags().GetStringSlice(flagParams)
+		skipProto, _       = cmd.Flags().GetBool(flagSkipProto)
 	)
 
 	if noDefaultModule && len(params) > 0 {
@@ -124,6 +126,7 @@ func scaffoldChainHandler(cmd *cobra.Command, args []string) error {
 		addressPrefix,
 		noDefaultModule,
 		skipGit,
+		skipProto,
 		minimal,
 		isConsumer,
 		params,